Codebase cleanup: extract dashboard, fix dead code, update README
- Extract dashboard rendering into dashboard.py (143 lines)
  - status_table(), status_json(), status_watch() with shared _build_table()
  - cli.py reduced from 757 to 580 lines
- Remove dead code: cmd_start had duplicated inline tmux launch logic
  that was left behind when _launch_in_tmux was introduced
- Remove unused import: time no longer needed in cli.py
- Update CLI module docstring to list all current subcommands
- Update README:
  - Document new commands: restart, attach, next, status --watch/--json
  - Add Session Management section (concurrency, branch detection, STOPPED)
  - Add tmux integration recipes
  - Update prerequisites to mention rich dependency
  - Organize commands into Task lifecycle / Monitoring / Navigation groups
